refactor: avoid repetition
