+++ client.go
@@ -2,8 +2,10 @@ package monta
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 	"net/http"
+	"net/url"
 	"time"
 )
 
@@ -106,4 +108,40 @@ func (c *Client) getToken(ctx context.Context) (_ *Token, err error) {
 	}
 	c.token = createdToken
 	return createdToken, nil
+}
+
+// Calls the given path and decode the response into the given type
+func getEntity[T any](ctx context.Context, client *Client, path string) (_ *T, err error) {
+	method := http.MethodGet
+	defer func() {
+		if err != nil {
+			err = fmt.Errorf("%s %s: %w", method, path, err)
+		}
+	}()
+	requestURL, err := url.Parse(apiHost + path)
+	if err != nil {
+		return nil, err
+	}
+	httpRequest, err := http.NewRequestWithContext(ctx, method, requestURL.String(), nil)
+	if err != nil {
+		return nil, err
+	}
+	if err := client.setAuthorization(ctx, httpRequest); err != nil {
+		return nil, err
+	}
+	httpResponse, err := client.httpClient.Do(httpRequest)
+	if err != nil {
+		return nil, err
+	}
+	defer func() {
+		_ = httpResponse.Body.Close()
+	}()
+	if httpResponse.StatusCode != http.StatusOK {
+		return nil, newStatusError(httpResponse)
+	}
+	var response T
+	if err := json.NewDecoder(httpResponse.Body).Decode(&response); err != nil {
+		return nil, err
+	}
+	return &response, nil
 }
